PHP Ported To
Mostly, there are a handful of things that I could do better, but for
now this is what I've got, so there. It kind of works. There are a lot
of dirty hacks and bodges that I'm not proud of. I'll clean up the code
later if I have time.